Inject `ProcessingManager` as a dependency
- A stub `ProcessingManager` is used for testing which will always
  return `true` for `isRepoReadyForProcessing` otherwise we can't test
  the processor without running gitea.
