+++ processor/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "dev": "nodemon src/server.js --watch src/",
     "start": "node src/server.js",
-    "test": "LOG_LEVEL=debug DATA_DIR=/data/test mocha test/integration --timeout=120000",
+    "test": "LOG_LEVEL=debug DATA_DIR=/data/test mocha test/integration --timeout=120000 --exit",
     "posttest": "rm -rf /data/test",
     "fmt": "prettier --write '*.json' $(find src/ test/ -name '*.js')"
   },
