+++ processor/src/app.js
@@ -2,11 +2,15 @@ const express = require('express')
 
 const { createProjectsAPI } = require('./projectAPI')
 const { createRemoteAPI } = require('./remoteAPI')
+const { ProcessingManager } = require('./processingManager')
 
-function createApp(repoDir = '/gitea-data/git/repositories') {
+function createApp(
+  repoDir = '/gitea-data/git/repositories',
+  processingManager = ProcessingManager,
+) {
   const app = express()
 
-  createProjectsAPI(app, repoDir)
+  createProjectsAPI(app, repoDir, processingManager)
   createRemoteAPI(app)
 
   return app
