+++ processor/src/watcher.js
@@ -1,33 +1,28 @@
-const { exponential: backOff, ExponentialOptions } = require('backoff')
 const chokidar = require('chokidar')
-const { Client } = require('pg')
 const debounce = require('lodash.debounce')
 const jsYaml = require('js-yaml')
 const log = require('loglevel')
 const path = require('path')
 
-const { DATA_DIR, GITEA_DB_CONFIG } = require('./env')
+const { DATA_DIR } = require('./env')
+const { ProcessingManager } = require('./processingManager')
 const { exists, exec, writeFile, readFile } = require('./utils')
 const processGerbers = require('./tasks/processGerbers')
 const processBOM = require('./tasks/processBOM')
 const processIBOM = require('./tasks/processIBOM')
 const processKicadPCB = require('./tasks/processKicadPCB')
 const processReadme = require('./tasks/processReadme')
-const { delay, reject } = require('lodash')
-
-const client = new Client(GITEA_DB_CONFIG)
-
-client.connect(e => {
-  if (e) {
-    log.error('Failed to connect to gitea DB')
-    throw e
-  } else {
-    log.info('Connected to gitea DB')
-  }
-})
 
 const running = {}
-function watch(events, repoDir = '/repositories') {
+
+/**
+ *
+ * @param {*} events
+ * @param {string=} repoDir
+ * @param {ProcessingManager} processingManager
+ * @returns
+ */
+function watch(events, repoDir = '/repositories', processingManager) {
   let dirWatchers = {}
 
   // watch repositories for file-system events and process the project
@@ -38,7 +33,10 @@ function watch(events, repoDir = '/repositories') {
     // additionally we ignore any invocations that happen while it's already running
     // to prevent it from trying to overwrite files that are already being written to
     const debouncedProcessRepo = debounce(async () => {
-      if ((await isProcessable(gitDir)) && !running[gitDir]) {
+      if (
+        (await processingManager.isRepoReadyForProcessing(gitDir)) &&
+        !running[gitDir]
+      ) {
         running[gitDir] = true
         await processRepo(events, repoDir, gitDir).catch(e => {
           log.error(`Error processing '${gitDir}':`, e)
@@ -73,7 +71,7 @@ function watch(events, repoDir = '/repositories') {
     }
     dirWatchers = {}
     watcher = chokidar.watch(repoWildcard).on('addDir', handleAddDir)
-  }, 60000)
+  }, 60_000)
 
   const unwatch = () => {
     clearInterval(timer)
@@ -145,197 +143,6 @@ async function getKitspaceYaml(checkoutDir) {
     ([yaml, yml, kitnicYaml, kitnicYml]) => yaml || yml || kitnicYaml || kitnicYml,
   )
   return jsYaml.safeLoad(yamlFile) || {}
-}
-
-/**
- *
- * @param {string} gitDir
- */
-async function isProcessable(gitDir) {
-  const { repoName, ownerName } = parseRepoGitDir(gitDir)
-  const { id, isEmpty, isMirror } = await queryGiteaRepoDetails(ownerName, repoName)
-
-  if (isEmpty) {
-    log.debug(
-      `Repo: ${ownerName}/${repoName} has no branches, it won't be processed!`,
-    )
-    return false
-  }
-
-  if (isMirror) {
-    // See gitea/models/task.go
-    // https://github.com/kitspace/gitea/blob/d1342b36f1b5e4ac0271d0f01c5270cd087f3575/models/task.go#L22
-    const MigrationStatusIsDone = 4
-    const status = await queryGiteaRepoMigrationStatus(id)
-
-    if (status === MigrationStatusIsDone) {
-      log.debug(
-        `Repo: ${ownerName}/${repoName} migration is done, the repo is processable.`,
-      )
-      return true
-    } else {
-      log.debug(
-        `Repo: ${ownerName}/${repoName} is ${status} not ${MigrationStatusIsDone}, the repo is unprocessable.`,
-      )
-      return false
-    }
-  }
-
-  log.debug(`Repo: ${ownerName}/${repoName} is processable.`)
-  return true
-}
-
-/**
- *
- * @param {string} gitDir the file-system path for the git repo
- * @returns {{ownerName: string, repoName: string}} repo details
- */
-function parseRepoGitDir(gitDir) {
-  const matcher =
-    /^\/gitea-data\/git\/repositories\/(?<ownerName>\w+)\/(?<repoName>\w+)/
-  const { ownerName, repoName } = gitDir.match(matcher)?.groups
-
-  if (ownerName == null || repoName == null) {
-    throw new Error(`Failed to parse gitDir: ${gitDir}`)
-  }
-
-  return { ownerName, repoName }
-}
-
-/**
- *
- * @param {string} ownerName
- * @param {string} repoName
- * @returns {Promise<{id: string, isEmpty: boolean, isMirror: boolean}>}
- */
-async function queryGiteaRepoDetails(ownerName, repoName) {
-  const repoQuery = {
-    name: 'fetch-repository',
-    text:
-      'select id, is_mirror, is_empty from repository' +
-      ' where lower(owner_name)=$1 and lower_name=$2',
-    values: [ownerName, repoName],
-  }
-
-  /**
-   * Query the migration task status for a gitea repo, with exponential backoff.
-   * @returns
-   */
-  const queryGiteaRepoWithBackoff = async repoQuery => {
-    const [ownerName, repoName] = repoQuery.values
-
-    const onBackoff = async (num, delay, resolve) => {
-      log.debug(
-        `Backoff started, querying repo ${ownerName}/${repoName}: ${num} ${delay}ms`,
-      )
-      const repoQueryResult = await client.query(repoQuery)
-
-      if (repoQueryResult.rows.length === 1) {
-        const { id, is_mirror, is_empty } = repoQueryResult.rows[0]
-        log.debug(`Repo: ${ownerName}/${repoName} was found`)
-        return resolve({ id, isEmpty: is_empty, isMirror: is_mirror })
-      }
-    }
-
-    const onFail = reject =>
-      reject(
-        new Error(
-          `Repo: ${ownerName}/${repoName} not found after ${MaximumRetries} trials.`,
-        ),
-      )
-
-    return asyncBackoff(onBackoff, onFail, 10)
-  }
-
-  const { id, isEmpty, isMirror } = await queryGiteaRepoWithBackoff(repoQuery)
-
-  return { id, isMirror, isEmpty }
-}
-
-/**
- *
- * @param {string} repoId
- * @returns {Promise<number>}
- */
-async function queryGiteaRepoMigrationStatus(repoId) {
-  // See gitea/models/task.go
-  // https://github.com/kitspace/gitea/blob/d1342b36f1b5e4ac0271d0f01c5270cd087f3575/models/task.go#L22
-  const MigrationTaskType = 0
-  const migrationStatusQuery = {
-    name: 'fetch-migration-status',
-    text: 'select status from task where repo_id=$1 and type=$2',
-    values: [repoId, MigrationTaskType],
-  }
-
-  /**
-   * Query the migration task status for a gitea repo, with exponential backoff.
-   * @returns
-   */
-  const queryMigrationStatusWithBackoff = async migrationStatusQuery => {
-    const onBackoff = async (num, delay, resolve) => {
-      log.debug(
-        'Backoff started, querying migration status for repo' +
-          `Id(${migrationStatusQuery.values[0]}): ${num} ${delay}ms`,
-      )
-
-      const migrationStatusQueryResult = await client.query(migrationStatusQuery)
-      if (migrationStatusQueryResult.rows.length === 1) {
-        const { status } = migrationStatusQueryResult.rows[0]
-        log.debug(`Repo: Id(${repoId})'s migration status: ${status}.`)
-        return resolve(status)
-      }
-    }
-
-    const onFail = reject =>
-      reject(
-        new Error(
-          `Repo: ${repoId} migration task not found after ${MaximumRetries} trials.`,
-        ),
-      )
-
-    return asyncBackoff(onBackoff, onFail, 10)
-  }
-
-  const repoMigrationStatus = await queryMigrationStatusWithBackoff(
-    migrationStatusQuery,
-  )
-
-  return repoMigrationStatus
-}
-
-/**
- *
- * @typedef {Object} ExponentialOptions
- * @property {number=} randomisationFactor
- * @property {number=} initialDelay
- * @property {number=} maxDelay
- * @property {number=} factor
- *
- * @param {function(number,  number, function): Promise<void>} onBackoff
- * @param {function(): Promise<void>} onFail
- * @param {number} maximumRetries
- * @param {ExponentialOptions=} config
- * @returns
- */
-async function asyncBackoff(onBackoff, onFail, maximumRetries, config) {
-  const queryBackoff = backOff(config)
-  queryBackoff.failAfter(maximumRetries)
-
-  return new Promise((resolve, reject) => {
-    const resetBackoffAndResolve = value => {
-      queryBackoff.reset()
-      resolve(value)
-    }
-
-    queryBackoff
-      .on(
-        'backoff',
-        async (num, delay) => await onBackoff(num, delay, resetBackoffAndResolve),
-      )
-      .on('ready', () => queryBackoff.backoff())
-      .on('fail', async () => await onFail(reject))
-      .backoff()
-  })
 }
 
 async function getGitHash(checkoutDir) {
