+++ processor/test/integration/test_remote.js
@@ -10,6 +10,7 @@ const { createApp } = require('../../src/app')
 
 const tmpDir = '/data/test/kitspace-processor-test-from-remote'
 const emptyRepoDir = '/data/test/kitspace-processor-test-empty-repo-dir'
+const processingManagerStub = { isRepoReadyForProcessing: async () => true }
 
 const REMOTE_API_TOKEN = process.env.REMOTE_API_TOKENS.split(',')
   .map(x => x.trim())
@@ -19,7 +20,7 @@ describe('remote API', () => {
   beforeEach(async () => {
     await exec(`mkdir -p ${tmpDir}`)
     await exec(`mkdir -p ${emptyRepoDir}`)
-    this.app = createApp(emptyRepoDir)
+    this.app = createApp(emptyRepoDir, processingManagerStub)
     this.supertest = supertest(this.app)
   })
